requirements.txt
playwright==1.41.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.1
